Refactor to organize classes
