+++ app/src/main/java/com/iamtechknow/worldview/picker/LayerPresenter.java
@@ -1,6 +1,5 @@
 package com.iamtechknow.worldview.picker;
 
-import com.iamtechknow.worldview.RxBus;
 import com.iamtechknow.worldview.model.Layer;
 
 import java.util.ArrayList;
