+++ app/src/main/java/com/iamtechknow/worldview/picker/NonLayerPresenter.java
@@ -1,7 +1,5 @@
 package com.iamtechknow.worldview.picker;
 
-import com.iamtechknow.worldview.RxBus;
-
 import java.util.ArrayList;
 import java.util.TreeMap;
